Add Render list on Products page
use context to make saved state
available to individual product
