+++ Week-1/src/context/ProductContext.js
@@ -1,0 +1,6 @@
+import { createContext } from 'react'
+const ProductContext = createContext({})
+const ProductProvider = ProductContext.Provider
+const ProductConsumer = ProductContext.Consumer
+
+export { ProductProvider, ProductConsumer}
